fastapi-backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database & Supabase (Fixed versions)
supabase==2.3.4
gotrue==1.3.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.24.0,<0.26.0
pydantic==2.5.3
pydantic[email]==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# CORS
starlette==0.35.1

# Payment Gateway
razorpay==1.4.1
